Extract path filtering login into PathCommitFilter to introduce as few changes a possible
